metrics: add prometheus format support for storage metrics
