+++ src/http_server/CMakeLists.txt
@@ -6,6 +6,7 @@ endif()
 set(src
   flb_hs.c
   flb_hs_endpoints.c
+  flb_hs_prometheus.c
   flb_hs_utils.c
   )
