+++ src/http_server/api/v1/metrics.h
@@ -25,6 +25,5 @@
 #include <fluent-bit/flb_sds.h>
 
 int api_v1_metrics(struct flb_hs *hs);
-flb_sds_t metrics_help_txt(char *metric_name, flb_sds_t *metric_helptxt);
 
 #endif
